requirements.txt
pandas
numpy
requests
osmnx
geopandas
shapely
tqdm
pyyaml
scipy
scikit-learn
transformers
torch
matplotlib
seaborn
wordcloud
streamlit
pydeck
groq
statsmodels
